[master] Handling errors in XMLFilesReaderManager for production jobs with invalid XML

See merge request lhcb-dirac/LHCbDIRAC!1243

+++ src/LHCbDIRAC/BookkeepingSystem/Service/XMLReader/XMLFilesReaderManager.py
@@ -589,7 +589,7 @@ class XMLFilesReaderManager(object):
 
         res = self.bkClient_.insertJob(attrList)
 
-        if not res["OK"] and production < 0:
+        if not res["OK"] and production is not None and production < 0:
             self.log.error("Failed inserting job", res["Message"])
             retVal = self.bkClient_.deleteProductionsContainer(production)
             if not retVal["OK"]:
